// frontend/src/types/generated/notifications/conectaPRONotificationsModule.schemas.ts
export type TemplateCategory = typeof TemplateCategory[keyof typeof TemplateCategory];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const TemplateCategory = {
  transactional: 'transactional',
  marketing: 'marketing',
  system: 'system',
  alert: 'alert',
  reminder: 'reminder',
  welcome: 'welcome',
  confirmation: 'confirmation',
  notification: 'notification',
  report: 'report',
  survey: 'survey',
} as const;

export type TemplateCreateDescription = string | null;

export type TemplateCreateChannelId = string | null;

export type TemplateCreateEmailSubject = string | null;

export type TemplateCreateEmailBodyHtml = string | null;

export type TemplateCreateEmailBodyText = string | null;

export type TemplateCreateEmailFromName = string | null;

export type TemplateCreateEmailFromAddress = string | null;

export type TemplateCreateEmailReplyTo = string | null;

export type TemplateCreateWhatsappTemplateName = string | null;

export type TemplateCreateWhatsappBody = string | null;

export type TemplateCreateWhatsappFooter = string | null;

export type TemplateCreateWhatsappButtonsItem = { [key: string]: unknown };

export type TemplateCreateSmsBody = string | null;

export type TemplateCreatePushTitle = string | null;

export type TemplateCreatePushBody = string | null;

export type TemplateCreatePushImageUrl = string | null;

export type TemplateCreatePushActionUrl = string | null;

export type TemplateCreateInAppTitle = string | null;

export type TemplateCreateInAppBody = string | null;

export type TemplateCreateInAppIcon = string | null;

export type TemplateCreateInAppActionUrl = string | null;

export type TemplateCreateSampleData = { [key: string]: unknown };

/**
 * Schema para criação de template.
 */
export interface TemplateCreate {
  /**
   * @minLength 1
   * @maxLength 200
   */
  name: string;
  /**
   * @minLength 1
   * @maxLength 200
   */
  slug: string;
  description?: TemplateCreateDescription;
  category?: TemplateCategory;
  status?: TemplateStatus;
  channel_id?: TemplateCreateChannelId;
  email_subject?: TemplateCreateEmailSubject;
  email_body_html?: TemplateCreateEmailBodyHtml;
  email_body_text?: TemplateCreateEmailBodyText;
  email_from_name?: TemplateCreateEmailFromName;
  email_from_address?: TemplateCreateEmailFromAddress;
  email_reply_to?: TemplateCreateEmailReplyTo;
  whatsapp_template_name?: TemplateCreateWhatsappTemplateName;
  whatsapp_body?: TemplateCreateWhatsappBody;
  whatsapp_footer?: TemplateCreateWhatsappFooter;
  whatsapp_buttons?: TemplateCreateWhatsappButtonsItem[];
  sms_body?: TemplateCreateSmsBody;
  push_title?: TemplateCreatePushTitle;
  push_body?: TemplateCreatePushBody;
  push_image_url?: TemplateCreatePushImageUrl;
  push_action_url?: TemplateCreatePushActionUrl;
  in_app_title?: TemplateCreateInAppTitle;
  in_app_body?: TemplateCreateInAppBody;
  in_app_icon?: TemplateCreateInAppIcon;
  in_app_action_url?: TemplateCreateInAppActionUrl;
  variables?: TemplateVariableSchema[];
  sample_data?: TemplateCreateSampleData;
  /** @maxLength 10 */
  locale?: string;
  requires_approval?: boolean;
  tags?: string[];
}

export type TemplateResponseDescription = string | null;

export type TemplateResponseChannelId = string | null;

export type TemplateResponseEmailSubject = string | null;

export type TemplateResponseWhatsappTemplateName = string | null;

export type TemplateResponseSmsBody = string | null;

export type TemplateResponsePushTitle = string | null;

export type TemplateResponseInAppTitle = string | null;

export type TemplateResponseVariablesItem = { [key: string]: unknown };

export type TemplateResponseOpenRate = number | null;

export type TemplateResponseClickRate = number | null;

export type TemplateResponseUpdatedAt = string | null;

/**
 * Schema de resposta para template.
 */
export interface TemplateResponse {
  /**
   * @minLength 1
   * @maxLength 200
   */
  name: string;
  /**
   * @minLength 1
   * @maxLength 200
   */
  slug: string;
  description?: TemplateResponseDescription;
  category?: TemplateCategory;
  status?: TemplateStatus;
  id: string;
  tenant_id: string;
  channel_id?: TemplateResponseChannelId;
  email_subject?: TemplateResponseEmailSubject;
  whatsapp_template_name?: TemplateResponseWhatsappTemplateName;
  sms_body?: TemplateResponseSmsBody;
  push_title?: TemplateResponsePushTitle;
  in_app_title?: TemplateResponseInAppTitle;
  variables?: TemplateResponseVariablesItem[];
  version?: number;
  total_sent?: number;
  total_opened?: number;
  total_clicked?: number;
  open_rate?: TemplateResponseOpenRate;
  click_rate?: TemplateResponseClickRate;
  active?: boolean;
  created_at: string;
  updated_at?: TemplateResponseUpdatedAt;
}

/**
 * Status do template.
 */
export type TemplateStatus = typeof TemplateStatus[keyof typeof TemplateStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const TemplateStatus = {
  draft: 'draft',
  pending_approval: 'pending_approval',
  approved: 'approved',
  rejected: 'rejected',
  active: 'active',
  inactive: 'inactive',
  archived: 'archived',
} as const;

export type TemplateUpdateName = string | null;

export type TemplateUpdateDescription = string | null;

export type TemplateUpdateCategory = TemplateCategory | null;

export type TemplateUpdateStatus = TemplateStatus | null;

export type TemplateUpdateEmailSubject = string | null;

export type TemplateUpdateEmailBodyHtml = string | null;

export type TemplateUpdateEmailBodyText = string | null;

export type TemplateUpdateWhatsappBody = string | null;

export type TemplateUpdateSmsBody = string | null;

export type TemplateUpdatePushTitle = string | null;

export type TemplateUpdatePushBody = string | null;

export type TemplateUpdateInAppTitle = string | null;

export type TemplateUpdateInAppBody = string | null;

export type TemplateUpdateVariables = TemplateVariableSchema[] | null;

export type TemplateUpdateSampleDataAnyOf = { [key: string]: unknown };

export type TemplateUpdateSampleData = TemplateUpdateSampleDataAnyOf | null;

export type TemplateUpdateActive = boolean | null;

/**
 * Schema para atualização de template.
 */
export interface TemplateUpdate {
  name?: TemplateUpdateName;
  description?: TemplateUpdateDescription;
  category?: TemplateUpdateCategory;
  status?: TemplateUpdateStatus;
  email_subject?: TemplateUpdateEmailSubject;
  email_body_html?: TemplateUpdateEmailBodyHtml;
  email_body_text?: TemplateUpdateEmailBodyText;
  whatsapp_body?: TemplateUpdateWhatsappBody;
  sms_body?: TemplateUpdateSmsBody;
  push_title?: TemplateUpdatePushTitle;
  push_body?: TemplateUpdatePushBody;
  in_app_title?: TemplateUpdateInAppTitle;
  in_app_body?: TemplateUpdateInAppBody;
  variables?: TemplateUpdateVariables;
  sample_data?: TemplateUpdateSampleData;
  active?: TemplateUpdateActive;
}

export type TemplateVariableSchemaDefault = unknown | null;

export type TemplateVariableSchemaDescription = string | null;

/**
 * Schema para variável de template.
 */
export interface TemplateVariableSchema {
  /**
   * @minLength 1
   * @maxLength 50
   */
  name: string;
  type?: string;
  required?: boolean;
  default?: TemplateVariableSchemaDefault;
  description?: TemplateVariableSchemaDescription;
}